LayoutTests/canvas/philip/tests/2d.imageData.create1.initial-expected.txt

< [index] >
2d.imageData.create1.initial
createImageData(imgdata) returns transparent black data of the right size
References: 2d.imageData.create1.size 2d.imageData.create.initial 2d.imageData.initial Bugzilla
Actual output:
Passed
